Safari-609.1.11.1/LayoutTests/media/encrypted-media/clearKey/clearKey-cenc-audio-playback-mse-expected.txt


EME API is supported OK
Media source is opened OK
EVENT(encrypted)
MediaKeys is created OK
EVENT(message)
EVENT(keystatuseschange)
Session: keyId=b0b1b2b3b4b5b6b7b8b9babbbcbdbebf status=usable OK
EXPECTED (video.currentTime >= '2') OK
END OF TEST
